Add test for Test Api
