+++ lib/gmo_payment/api_errors.rb
