+++ lib/gmo_payment/version.rb
@@ -1,3 +1,3 @@
 module GmoPayment
-  VERSION = '0.0.2'
+  VERSION = '0.0.2'.freeze
 end
