--- test/test_gmo_payment.rb
@@ -1,16 +1,0 @@
-require 'test/unit'
-require 'gmo_payment'
-require 'helper'
-
-class TestGmoPayment < Test::Unit::TestCase
-  def test_public_methods
-    set_basic_config
-    gmo_payment = GmoPayment::Api.new
-    api_methods = gmo_payment.public_methods(false)
-    stub_request_to_payment
-    api_methods.each do |api_call|
-      response = gmo_payment.public_send(api_call)
-      assert_equal(GmoPayment::ApiResponse, response.class)
-    end
-  end
-end
